refactor folder and files structure
