+++ .env.example
@@ -1,0 +1,1 @@
+BACKEND_URL=your backend server url
