+++ README.md
@@ -1,1 +1,5 @@
 # quickss-web-app
+
+helpers in UI
+folders fo components
+naming routes path
