+++ package.json
@@ -15,6 +15,7 @@
     "@types/react-redux": "^7.1.11",
     "@types/react-router-dom": "^5.1.6",
     "axios": "^0.21.1",
+    "dotenv": "^8.2.0",
     "google-maps-react": "^2.0.6",
     "material-ui-dropzone": "^3.5.0",
     "node-sass": "4.14.1",
