+++ src/components/Admin/ClientRegistration/Upload/styles.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Styles = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default Styles
