+++ src/components/Client/ProductRegistration/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const index = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default index
