+++ src/components/Client/ProductRegistration/styles.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const styles = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default styles
