+++ src/containers/Client/Dashboard.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Dashboard = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default Dashboard
