+++ src/store/auth/actions.ts
@@ -14,7 +14,7 @@ export const authActions = (
       dispatchHandler({ type: LOGIN, data: data, dispatch });
       localStorage.setItem("QUICKSS-USER-TOKEN", data.token);
       localStorage.setItem("QUICKSS-USER-ROLE", data.user.role);
-      window.location.href = "/view-client";
+      window.location.href = "/admin/dashboard";
     }
   } catch (error) {
     if (error) {
