+++ src/store/rootReducer.ts
@@ -1,5 +1,5 @@
 import { combineReducers } from "redux";
-import { hotelRestoReducer } from "./client/reducers";
+import { hotelRestoReducer } from "./admin/reducers";
 import { authReducer } from "./auth/reducers";
 
 export const rootReducer = combineReducers({
